Refactor process data
