+++ nerfstudio/utils/scripts.py
@@ -1,0 +1,43 @@
+"""Helpers for running script commands."""
+# Copyright 2022 The Nerfstudio Team. All rights reserved.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+import subprocess
+import sys
+from typing import Optional
+
+from rich.console import Console
+
+CONSOLE = Console(width=120)
+
+
+def run_command(cmd: str, verbose=False) -> Optional[str]:
+    """Runs a command and returns the output.
+
+    Args:
+        cmd: Command to run.
+        verbose: If True, logs the output of the command.
+    Returns:
+        The output of the command if return_output is True, otherwise None.
+    """
+    out = subprocess.run(cmd, capture_output=not verbose, shell=True, check=False)
+    if out.returncode != 0:
+        CONSOLE.rule("[bold red] :skull: :skull: :skull: ERROR :skull: :skull: :skull: ", style="red")
+        CONSOLE.print(f"[bold red]Error running command: {cmd}")
+        CONSOLE.rule(style="red")
+        CONSOLE.print(out.stderr.decode("utf-8"))
+        sys.exit(1)
+    if out.stdout is not None:
+        return out.stdout.decode("utf-8")
+    return out
